requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
peewee==3.14.4
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
python-dotenv==0.16.0
six==1.15.0
Werkzeug==1.0.1
